[Data] Remove issue detector constructor (#58852)
## Description

Based on the comment here:
https://github.com/ray-project/ray/pull/58630#issuecomment-3550589069

Current `IssueDetector` base class requires all its subclasses include
the `StreamingExecutor` as the arguments, making classes hard to mock
and test because we have to mock all of StreamingExecutor.

In this PR, we did following:
1. Remove constructor in `IssueDetector` base class and add
`from_executor()` to setup the class based on the executor
2. Refactor subclasses of `IssueDetector` to use this format

## Related issues

Related to https://github.com/ray-project/ray/issues/58562

## Additional information
> Optional: Add implementation details, API changes, usage examples,
screenshots, etc.

---------

Signed-off-by: machichima <nary12321@gmail.com>