[core] Fix crash when using jvm HDFS by adding RAY_DISABLE_FAILURE_SIGNAL_HANDLER
Ray's Abseil failure signal handler conflicts with JVM's internal signal handling. When PyArrow's HadoopFileSystem starts a JVM, and then Ray installs its signal handler, JVM's normal internal SIGSEGV operations (for safepoint synchronization) trigger Ray's handler, causing a crash.

It seems we already have some code which disables this for java bindings
(there's a todo about in io_ray_runtime_RayNativeRuntime.cc). This is
all well and good when we realize we're running in a java runtime, but
for the bug report I'm not sure how obvious it is (maybe we should
consider doing something like this in data itself automatically?)
Unsure.  For our part if we know the datasource is form hdfs, that seems
just as likely as any that this might be happening. But it's only
relevant if you're running the jvm in the same process (I was only able
to reproduce this by running everything in the same process).  One of
the comments in the bug ticket points out they're able to skirt the
error by doing process isolaiton (which ended up being the hint I
needed).

Signed-off-by: zac <zac@anyscale.com>