[train] Add bundle_label_selector to ScalingConfig (#58845)
This PR adds a `bundle_label_selector` argument to the `ScalingConfig`
that allows Ray Train workers to be placed on nodes with particular
labels. The previous workaround, namely using `resources_per_worker`, is
less flexible.

`bundle_label_selector` can either be a single dict, in which case it
will apply to all the workers, or a list of length `num_workers`, in
which case each item in the list will correspond to one of the workers.

I added verification to the controller instead of validating that none
of the callbacks have `on_controller_start_worker_group` when
`bundle_label_selector` is set because we might change
`on_controller_start_worker_group` in the future. We can revisit this
issue then.

---------

Signed-off-by: Timothy Seah <tseah@anyscale.com>