[RLlib] Error out when not being able to load data (#59002)
## Description

Main scope of this PR:
If there is no data at `data_path`, we currently don't error out but
just log a warning and continue.
So the error that user gets is some thing further down the line.
This PR makes it so that we error out if the data does not exist - there
is no reason for us to mask that and try to continue still.

Secondary scope:
We introduced some formatting for log messages that looks to me like it
adhere to some standard but I can not find that format anywhere else in
Ray. This PR removes this formatting to help us to not creep into a
veriety of such formats across our codebase.

Signed-off-by: CesarRodrigu <128821412+CesarRodrigu@users.noreply.github.com>