[core] remove cluster_full_of_actors_detected_* as autoscaler v2 doesn't use them (#59052)
This PR removes the `cluster_full_of_actors_detected` and
`cluster_full_of_actors_detected_by_gcs` fields from the protobuf, as
they are not used in autoscaler v2, and autoscaler v1 is scheduled for
deletion soon.

The 2 fields are considered private, so they are deleted without
maintaining backward compatibility.

Signed-off-by: Rueian <rueiancsie@gmail.com>