[core] fix counter metric default branch (#59095)
`RAY_enable_open_telemetry` is now `True` (not `False`) by default. This
means that we need to swap these if-branch clause. The current
implementation means that if someone set
`RAY_enable_open_telemetry=True` manually (which users don't have to but
they might), it will go to the incorrect branch.

Test:
- CI

Signed-off-by: Cuong Nguyen <can@anyscale.com>