[core] change RAY_metric_cardinality_level default to recommended (#59096)
Change the default value of `RAY_metric_cardinality_level` to be
recommended. Since `recommended` drops the `worker_id` label in
`ray_tasks` and `ray_actors` metrics, some tests that rely on the
uniqueness of labels to distinguish samples fail. To fix it, I force
them to have unique task name, so that the sample can be uniquely
identified by label set.

Test:
- CI

---------

Signed-off-by: Cuong Nguyen <can@anyscale.com>