[core] Make CancelTask RPC Fault Tolerant (#58018)
> Briefly describe what this PR accomplishes and why it's needed.

Makes CancelTask RPC Fault Tolerant. Created an intermediary RPC similar
to what was done in #57648 in that when the force_exit flag is enabled
for cancel, the executor worker is shut down gracefully. However we have
no way of determining whether the shutdown was successful on the owner
core worker, hence we send the cancel request to the raylet via a new
RPC CancelLocalTask that guarantees the worker is killed. Added a python
test to verify retry behavior, leaving out the cpp test after talking to
@dayshah due to being a bit complicated in that we need to take into
account all orderings of the owner/executor states in the cancellation
process.

In the current task submission path, we don't keep track of the raylet
address of the worker when we receive the PushTask RPC. It's a bit
complicated to do this since the GCS keeps track of actor lifecycles
including requesting leases, hence instead of touching the hot path
(task submission) we decided to just complicate the cancellation path.
Upon receiving a CancelTask RPC, we will query the gcs node cache to get
the node info. Only if its not in the cache do we then query the GCS.
Unfortunately the gcs node cache is currently not thread safe and show
only be accessed on the main io service hence we refactored
Normal/ActorTaskSubmitter so that it posts the portion of the code that
accesses the cache onto the main io service.

There was also a race condition in CancelLocalTask/KillLocalActor where
send_reply_callback could be triggered twice if we receive the response
from CancelTask/KillActor, but the worker is not evicted from the raylet
worker pool immediately. Hence the callback in execute_after could
trigger. So added a replied boolean flag to guard against this.

CancelTask when force_kill=true behavior has been modified to trigger a
SIGKILL after a set amount of time if the graceful shutdown from the
worker hangs.

Lastly, the actor task retry did not use the config used in the normal
task retry timer. Updated this

---------

Signed-off-by: joshlee <joshlee@anyscale.com>
Co-authored-by: Edward Oakes <ed.nmi.oakes@gmail.com>
Co-authored-by: Jiajun Yao <jeromeyjj@gmail.com>