[Data] Fix reading from written parquet for numpy with NaNs (#59172)
## Description
Currently when transforming pandas to arrow block, we will convert it
into `pa.null` if all the value in TensorArray is NaN or empty string,
however all NaN might actually represent something and could actually
happen in transformation.


https://github.com/ray-project/ray/blob/8b4a1ee98fd92a972d07e9c08af19f02829dd40f/python/ray/air/util/tensor_extensions/pandas.py#L926-L958

And could triggered error when we tried to convert that arrow block back
to pandas (try running the repro script from issue)

This PR skip coerced to `pa.null` if the`col.dtype` is `TensorDtype`,
which should tolerate the all NaN or all empty string table.

## Related issues
Closes #59087

---------

Signed-off-by: You-Cheng Lin <mses010108@gmail.com>
Signed-off-by: You-Cheng Lin <106612301+owenowenisme@users.noreply.github.com>
Co-authored-by: Alexey Kudinkin <alexey.kudinkin@gmail.com>