[RLlib] MetricsLogger  tweaks+ Stats rewrite (#56838)
## Why are these changes needed?

In this PR, we refactor MetricsLogger and Stats into a design that is
more sustainable (more tests + ability for users to modify behaviour
through configuration) and also improves on a number of key issues.

**The main issues with the old design are:**

**Too many useless combinations of input arguments to methods like
MetricsLogger.log_value()**
These increase the test surface area a lot and is solved by refactoring
the legacy Stats class into multiple stats classes such as StatsBase,
MaxStats or LifetimeSumStats. These classes represent the few valid
combinations of input arguments such as (reduce="mean", window=10).

**Many metrics are easily skewed because windows are applied in parallel
components AND in the root metrics logger**
This is opaque and not easy to work around. This PR solves this by
treating stats at the root metrics logger differently (stats have
variables "is_root" and "is_leaf" to align their behaviour when pushing,
reducing, peeking or merging values with where they are. For example,
stats in the root metrics logger are mostly used to aggregate metrics
and therefore we don't apply a window, even if the values at leafs are
logged with a window.
This also has the nice sideeffect that we don't need to specify how to
prioritize values when merging anymore (see the deprecated
"reduce_per_index_on_aggregate" argument.

**Different modes of aggregates (sometimes we use MetricsLogger.log_dict
and sometimes MetricsLogger.aggregate)**
This leads to different aggregation patters. For example, we use
MetricsLogger.log_dict() to log results from Learner updates. This can
happen n times per training iteration and stats with a window of 1 are
overridden each call such that we report only the last value logged
instead of actual averages.

**Skewed throughputs**
Before this PR, throughputs are calculated by calculating the
exponential weighted average over increments divided by the time passed
since the last update for the given metric. This can lead to a noticable
bias although, anecdotally, we found that some throughput metrics where
off by a few % in the cases we looked at.
This PR improves on this by calculating all throughputs by dividing the
sum of all increments per iteration by the total iteration time.

Other notable changes:
- No need for "tensor mode". Keep all tensors as they are, detach and
move to CPU as soon as we peek, reduce or create a state. Therefore,
when instantiating, all values are on CPU and implicitly converted to
Tensors and moved to GPU as soon as values are logged.
- Discriminate between Stats being leaf/non-leaf and root/non-root. Leaf
stats can be pushed to. Non-leaf stats are used to aggregate. Root stats
can have a special function (like with LifetimeSum)
- Discriminate between MetricsLoggers being root and non-root to inform
Stats objects whether they are instantiated inside a root MetricsLogger.

This PR includes a mini migration guide that points out what needs to
happen from the users perspective.

---------

Signed-off-by: Artur Niederfahrenhorst <attaismyname@googlemail.com>
Co-authored-by: Daniel Sperber <github.blurry@9ox.net>