[Data] Close threads from async UDF after actor died (#59261)
## Description

When using async map with actors in Ray Data, the asyncio thread can
keep the actor process alive and block shutdown even after the actor is
marked as dead. This occurs because the asyncio thread does not exit
when the main thread finishes.

To fix this, we create the thread with `daemon=True`, ensuring it
terminates when the main thread exits.

## Related issues
Closes #59033

Signed-off-by: You-Cheng Lin <mses010108@gmail.com>