[rllib] Add `pytest_runtest_makereport` to all tests (#59003)
## Description
When a test fails in RLlib, we need to read through the logs to identify
the problem which are removed after sometime.
`pytest_runtest_makereport` exists to help generate a zip report of the
problem to allow easier offline problem solving.

We can't add this to our learning tests as they aren't pytest but normal
scripts

---------

Signed-off-by: Mark Towers <mark@anyscale.com>
Signed-off-by: Kamil Kaczmarek <kamil@anyscale.com>
Co-authored-by: Mark Towers <mark@anyscale.com>
Co-authored-by: Kamil Kaczmarek <kaczmarek.poczta@gmail.com>
Co-authored-by: Kamil Kaczmarek <kamil@anyscale.com>