[Core] Lift node ip resolution before init_temp to allow temp_dir to be fetched for the right node (#59229)
We are adding support for the ability to specify `temp_dir` per node in
https://github.com/ray-project/ray/pull/57735. However, this introduces
a problem when we attempt to connect to a node with a custom ip and temp
dir. Since the temp dir information will be stored on the `node_info`,
we will need to fetch the `node_info` of the node we are connecting to
from the gcs. However, since we are still initializing ray, the only
information we have access to is the node's ip address. If the node we
are connecting to has a custom ip address as well, there will be no way
for us to fetch the correct `node_info` needed to set up the temp dir.

To resolve this issue, this PR will resolve the node's IP via looking
what what IP was used to start the raylet (instead of the existing
implementation which depends on reading a file stored in a `temp_dir`
that hasn't been initialized yet). This way, we will always be able to
retrieve the `node_info` when connecting to even a node with a custom ip
address.

## Related issues
https://github.com/ray-project/ray/pull/57735

---------

Signed-off-by: davik <davik@anyscale.com>
Co-authored-by: davik <davik@anyscale.com>