[data] Update streaming_repartition and map_batches_fusion (#59476)
Analysis of the two operator patterns:

## Streaming_repartition → map_batches

| | Number of `map_batches` tasks |

|----------------------|---------------------------------------------------------------------------|
| **Fused** | `num_input_blocks` (which is ≤ number of output blocks of
StreamingRepartition) |
| **Not fused** | number of output blocks of StreamingRepartition |

When fused, the number of tasks equals the number of input blocks, which
is
≤ the number of output blocks of StreamingRepartition. If
StreamingRepartition
is supposed to break down blocks to increase parallelism, that won't
happen
when fused. So we don't fuse.

---

## Map_batches → streaming_repartition

`batch_size % target_num_rows == 0`

|                      | Number of `map_batches` tasks |
|----------------------|-------------------------------|
| **Fused**            | == total_rows / batch_size |
| **Not fused**        | == total_rows / batch_size |

So, the fusion doesn’t affect the parallelism.

---

Thus, we currently disable the `Streaming_repartition → map_batches`
fusion and enable the fusion when `batch_size % target_num_rows == 0`
for `Map_batches → streaming_repartition`.

---------

Signed-off-by: Xinyuan <43737116+xinyuangui2@users.noreply.github.com>
Signed-off-by: xgui <xgui@anyscale.com>
Signed-off-by: Alexey Kudinkin <alexey.kudinkin@gmail.com>
Co-authored-by: Alexey Kudinkin <alexey.kudinkin@gmail.com>