[RLlib] When retrying EnvRunner._sample, replace recursive with iterative solution (#56082)
## Why are these changes needed?

When we retry EnvRunner._sample, we call EnvRunner._sample again. This
makes for a recursive solution, leaving already done episodes in
done_episodes_to_return.

This PR makes it so that  ....

- We stay true to the promise of the docstring of the EnvRunner.sample()
methods that we return AT least n timesteps. Even if envs are restarted.
Before, we would not reset how many timessteps have been sampled when
resetting the environment (and thus starting a new episode to collect
chunks from).
- ... `done_episodes_to_return` does not start from zero if we retry in
the new recursive call
- ... we can make an arbitrary number of calls to env_runner.sample(),
sampling episodes and timesteps as we like. Before We'd break if the
number of episodes or timesteps was reached, thus leaving other episodes
in a dangling state even if they were finished. (run included test on
old version of code to confirm)
- ... we don't recurse, thereby creating more risk for future memory
leaks

---------

Co-authored-by: Kamil Kaczmarek <kaczmarek.poczta@gmail.com>
Signed-off-by: jasonwrwang <jasonwrwang@tencent.com>