[Data] Change `PhysicalOperator.completed()` to `has_completed()` (#59234)
## Description

This is a follow-up PR after: #58915 for changing the function name from
```completed()``` to ```has_completed()``` in the class
PhycicalOperator. This is in-line with our discussions with @bveeramani
here:
https://github.com/ray-project/ray/pull/58915#pullrequestreview-3497529812.
This is done to make it clear that the method doesn't modify state and
has no side effects.

## Related issues
#58884

## Additional information
Just updated the completed() and all it's references in other classes
and tests.

---------

Signed-off-by: Simeet Nayan <simeetnayan.8100@gmail.com>
Co-authored-by: Balaji Veeramani <balaji@anyscale.com>
Signed-off-by: jasonwrwang <jasonwrwang@tencent.com>